#25 refactor : user 관련 패키지 구조 변경 - userValidator 생성 - 구현하지 않은 부분 삭제
