--- be/issue/src/main/java/codesquad/issueTracker/comment/repository/CommentRepository.java
@@ -1,13 +1,0 @@
-package codesquad.issueTracker.comment.repository;
-
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class CommentRepository {
-	private final NamedParameterJdbcTemplate jdbcTemplate;
-
-	public CommentRepository(NamedParameterJdbcTemplate jdbcTemplate) {
-		this.jdbcTemplate = jdbcTemplate;
-	}
-}
