--- be/issue/src/main/java/codesquad/issueTracker/comment/service/CommentService.java
@@ -1,12 +1,0 @@
-package codesquad.issueTracker.comment.service;
-
-import org.springframework.stereotype.Service;
-
-import codesquad.issueTracker.comment.repository.CommentRepository;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Service
-public class CommentService {
-	private final CommentRepository commentRepository;
-}
