+++ be/issue/src/main/java/codesquad/issueTracker/global/config/OauthConfig.java
@@ -1,25 +1,27 @@
 package codesquad.issueTracker.global.config;
 
-import codesquad.issueTracker.jwt.domain.OAuthProperties;
-import codesquad.issueTracker.jwt.util.InMemoryProviderRepository;
-import codesquad.issueTracker.jwt.util.OauthAdapter;
-import codesquad.issueTracker.jwt.util.OauthProvider;
 import java.util.Map;
-import lombok.RequiredArgsConstructor;
+
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import codesquad.issueTracker.jwt.util.InMemoryProviderRepository;
+import codesquad.issueTracker.oauth.domain.OAuthProperties;
+import codesquad.issueTracker.oauth.util.OauthAdapter;
+import codesquad.issueTracker.oauth.util.OauthProvider;
+import lombok.RequiredArgsConstructor;
 
 @Configuration
 @EnableConfigurationProperties(OAuthProperties.class)
 @RequiredArgsConstructor
 public class OauthConfig {
 
-    private final OAuthProperties properties;
+	private final OAuthProperties properties;
 
-    @Bean
-    public InMemoryProviderRepository inMemoryProviderRepository() {
-        Map<String, OauthProvider> providers = OauthAdapter.getOauthProviders(properties);
-        return new InMemoryProviderRepository(providers);
-    }
+	@Bean
+	public InMemoryProviderRepository inMemoryProviderRepository() {
+		Map<String, OauthProvider> providers = OauthAdapter.getOauthProviders(properties);
+		return new InMemoryProviderRepository(providers);
+	}
 }
