--- be/issue/src/main/java/codesquad/issueTracker/issue/controller/IssueController.java
@@ -1,12 +1,0 @@
-package codesquad.issueTracker.issue.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-import codesquad.issueTracker.issue.service.IssueService;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@RestController
-public class IssueController {
-	private final IssueService issueService;
-}
