--- be/issue/src/main/java/codesquad/issueTracker/issue/repository/IssueRepository.java
@@ -1,13 +1,0 @@
-package codesquad.issueTracker.issue.repository;
-
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class IssueRepository {
-	private final NamedParameterJdbcTemplate jdbcTemplate;
-
-	public IssueRepository(NamedParameterJdbcTemplate jdbcTemplate) {
-		this.jdbcTemplate = jdbcTemplate;
-	}
-}
