--- be/issue/src/main/java/codesquad/issueTracker/issue/service/IssueService.java
@@ -1,12 +1,0 @@
-package codesquad.issueTracker.issue.service;
-
-import org.springframework.stereotype.Service;
-
-import codesquad.issueTracker.issue.repository.IssueRepository;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Service
-public class IssueService {
-	private final IssueRepository issueRepository;
-}
