+++ be/issue/src/main/java/codesquad/issueTracker/jwt/util/InMemoryProviderRepository.java
@@ -1,13 +1,15 @@
 package codesquad.issueTracker.jwt.util;
 
 import java.util.Map;
+
+import codesquad.issueTracker.oauth.util.OauthProvider;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class InMemoryProviderRepository {
-    private final Map<String, OauthProvider> providers;
+	private final Map<String, OauthProvider> providers;
 
-    public OauthProvider findByProviderName(String name) {
-        return providers.get(name);
-    }
+	public OauthProvider findByProviderName(String name) {
+		return providers.get(name);
+	}
 }
