--- be/issue/src/main/java/codesquad/issueTracker/jwt/util/OauthProvider.java
@@ -1,28 +1,0 @@
-package codesquad.issueTracker.jwt.util;
-
-import codesquad.issueTracker.jwt.domain.OAuthProperties;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-public class OauthProvider {
-    private final String clientId;
-    private final String clientSecret;
-    private final String redirectUrl;
-    private final String tokenUrl;
-    private final String userInfoUrl;
-
-    public OauthProvider(OAuthProperties.Client client, OAuthProperties.Provider provider) {
-        this(client.getClientId(), client.getClientSecret(), client.getRedirectUrl(), provider.getTokenUrl(), provider.getUserInfoUrl());
-    }
-
-    @Builder
-    public OauthProvider(String clientId, String clientSecret, String redirectUrl, String tokenUrl, String userInfoUrl) {
-        this.clientId = clientId;
-        this.clientSecret = clientSecret;
-        this.redirectUrl = redirectUrl;
-        this.tokenUrl = tokenUrl;
-        this.userInfoUrl = userInfoUrl;
-    }
-}
