--- be/issue/src/main/java/codesquad/issueTracker/label/repository/LabelRepository.java
@@ -1,13 +1,0 @@
-package codesquad.issueTracker.label.repository;
-
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class LabelRepository {
-	private final NamedParameterJdbcTemplate jdbcTemplate;
-
-	public LabelRepository(NamedParameterJdbcTemplate jdbcTemplate) {
-		this.jdbcTemplate = jdbcTemplate;
-	}
-}
