--- be/issue/src/main/java/codesquad/issueTracker/milestone/service/MilestoneService.java
@@ -1,12 +1,0 @@
-package codesquad.issueTracker.milestone.service;
-
-import org.springframework.stereotype.Service;
-
-import codesquad.issueTracker.milestone.repository.MilestoneRepository;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Service
-public class MilestoneService {
-	private final MilestoneRepository milestoneRepository;
-}
