+++ be/issue/src/main/java/codesquad/issueTracker/user/domain/User.java
@@ -1,5 +1,10 @@
 package codesquad.issueTracker.user.domain;
 
+import org.mindrot.jbcrypt.BCrypt;
+
+import codesquad.issueTracker.global.exception.CustomException;
+import codesquad.issueTracker.global.exception.ErrorCode;
+import codesquad.issueTracker.user.dto.LoginRequestDto;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -22,6 +27,16 @@ public class User {
 		this.profileImg = profileImg;
 		this.name = name;
 		this.loginType = loginType;
+	}
+
+	public void validateLoginUser(LoginRequestDto loginRequestDto) {
+		if (password == null) {
+			throw new CustomException(ErrorCode.GITHUB_LOGIN_USER);
+		}
+		if (!loginRequestDto.getEmail().equals(email)
+			|| !BCrypt.checkpw(loginRequestDto.getPassword(), password)) {
+			throw new CustomException(ErrorCode.FAILED_LOGIN_USER);
+		}
 	}
 
 }
